refactor(views): Refactor views
move methods into utils or services
remove unused code
refactor subprocesses

update blackletter
