+++ scanning/urls.py
@@ -15,7 +15,7 @@ from scanning.views import (
     compute_redactions_api,
     delete_page,
     dismiss_issue,
-    dismiss_issues,
+    # dismiss_issues,
     export_pdf,
     flag_issue,
     generate_files,
@@ -31,9 +31,9 @@ from scanning.views import (
     queue_view,
     recalculate,
     remove_flag,
-    repare_opinions,
+    # repare_opinions,
     reprocess,
-    reprocess_section_view,
+    # reprocess_section_view,
     scan_detail,
     scan_list,
     scan_process_view,
@@ -74,7 +74,7 @@ urlpatterns = [
     path("scans/<int:pk>/delete-page/", delete_page, name="delete_page"),
     path("scans/<int:pk>/insert/", add_page_insert, name="add_page_insert"),
     path("scans/<int:pk>/dismiss-issue/", dismiss_issue, name="dismiss_issue"),
-    path("scans/<int:pk>/dismiss/", dismiss_issues, name="dismiss_issues"),
+    # path("scans/<int:pk>/dismiss/", dismiss_issues, name="dismiss_issues"),
     path("scans/<int:pk>/detections/", serve_detections, name="serve_detections"),
     path("scans/<int:pk>/opinions-json/", serve_opinions, name="serve_opinions"),
     path("scans/<int:pk>/margin-rects/", serve_margin_rects, name="serve_margin_rects"),
@@ -93,8 +93,8 @@ urlpatterns = [
     path("scans/<int:pk>/flag/", flag_issue, name="flag_issue"),
     path("scans/<int:pk>/flag/<int:flag_id>/delete/", remove_flag, name="remove_flag"),
     path("scans/<int:pk>/add-single-detection/", add_single_detection, name="add_single_detection"),
-    path("scans/<int:pk>/repare-opinions/", repare_opinions, name="repare_opinions"),
-    path("scans/<int:pk>/reprocess-section/", reprocess_section_view, name="reprocess_section"),
+    # path("scans/<int:pk>/repare-opinions/", repare_opinions, name="repare_opinions"),
+    # path("scans/<int:pk>/reprocess-section/", reprocess_section_view, name="reprocess_section"),
     path("scans/<int:pk>/bake-redactions/", bake_redactions, name="bake_redactions"),
     path("scans/<int:pk>/export/", export_pdf, name="export_pdf"),
     path("opinions/", opinion_list, name="opinion_list"),
